docs: add example config
Initially define the Hammock Definition Language (HDL) as a subset of
TOML.

The goal here is to declaratively "program" the intended behaviour of a
given device. Hammock then executes this program to implement the
behaviour through event driven programming. Actions are performed by
matching rules, events are propagated to apps that support them, etc.

The HDL allows for defining match rules - the requirements and
properties of a cgroup which an application can be put in. As well as
properties of hammock's events and tags.

The match rules and events are currently all "magic" words which will
map to the Rust code. HDL is not actually turing complete and doesn't
support "building" events based on others (yet...). This would be nice
for the "work-ready" event for example, as this is really a compound of
a few unrelated things (we're on charge, unmetered network, idle, etc).
Being able to define this rule in HDL would make hammock pretty flexible
indeed....