It's all coming together now
The dbus API doesn't work for everything, some apps (ahem KGX) ignore
it, likely many many others. We'll need to polyfill
gnome_start_systemd_scope() to do this properly on pmOS and integrate
with systemd scopes for systemd distros.

The DBUS session can only be connected to when running hammock as
non-root, some perms stuff. It may be sensible to run a daemon as the
user and use something like ipc_channel to generate dbus and wayland
events.

cgroup stuff doesn't seem to work yet, assuming it's a permission error
that isn't getting bubbled up.